requirements.txt
click==6.6
pycrypto==2.6.1
six==1.10.0
